Add vault token-helper compatibility to `destruct store`
vault cli uses a token helper (i.e. stores a token in a file) when using
vault cli to login. instead of relying on the vault client automatically
using $VAULT_TOKEN as the token to create the destruct token, i added a token string input that defaults to either ~/.vault-token or $VAULT_ADDR